# Histogram performance: optimize floatBucketIterator, part 2